Удалил репозиторий деталей вакансии. Откорректировал интерактор деталей вакансии
