+++ app/src/main/java/ru/practicum/android/diploma/vacancy/details/di/VacancyDetailsDI.kt
@@ -3,20 +3,19 @@ package ru.practicum.android.diploma.vacancy.details.di
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 import ru.practicum.android.diploma.favorites.vacancies.domain.api.FavoritesVacanciesInteractor
-import ru.practicum.android.diploma.vacancy.details.data.VacancyDetailsRepositoryImpl
+import ru.practicum.android.diploma.search.domain.api.SearchInteractor
 import ru.practicum.android.diploma.vacancy.details.domain.api.VacancyDetailsInteractor
-import ru.practicum.android.diploma.vacancy.details.domain.api.VacancyDetailsRepository
 import ru.practicum.android.diploma.vacancy.details.domain.impl.VacancyDetailsInteractorImpl
 import ru.practicum.android.diploma.vacancy.details.presentation.viewmodel.VacancyDetailsViewModel
 
 val vacancyDetailsDomainModule = module {
 
-    single<VacancyDetailsRepository> {
-        VacancyDetailsRepositoryImpl(get(), get(), get<FavoritesVacanciesInteractor>())
-    }
-
     single<VacancyDetailsInteractor> {
-        VacancyDetailsInteractorImpl(get())
+        VacancyDetailsInteractorImpl(
+            searchInteractor = get<SearchInteractor>(),
+            favoritesInteractor = get<FavoritesVacanciesInteractor>(),
+//            dtoMapper = get()
+        )
     }
 }
