+++ app/src/main/java/ru/practicum/android/diploma/vacancy/details/presentation/viewmodel/VacancyDetailsViewModel.kt
@@ -6,7 +6,6 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-import retrofit2.HttpException
 import ru.practicum.android.diploma.favorites.vacancies.domain.api.FavoritesVacanciesInteractor
 import ru.practicum.android.diploma.search.domain.model.Address
 import ru.practicum.android.diploma.search.domain.model.Contacts
@@ -85,31 +84,12 @@ class VacancyDetailsViewModel(
 
                         is Result.Error -> {
                             _isLoading.value = false
-
-                            // Если переход из избранного и вакансия не найдена (404), удаляем из БД
-                            if (source == VacancyDetailsSource.FAVORITES &&
-                                isNotFoundError(result.throwable)
-                            ) {
-                                favoritesInteractor.removeFromFavorites(id)
-                            }
-
                             _error.value = result.throwable
                         }
                     }
                 }
         }
     }
-}
-
-private const val HTTP_NOT_FOUND = 404
-private fun isNotFoundError(throwable: Throwable?): Boolean {
-    val httpException = when (throwable) {
-        is HttpException -> throwable
-        else -> throwable?.cause as? HttpException
-    }
-
-    return throwable?.message == "404 Not Found" ||
-        httpException?.code() == HTTP_NOT_FOUND
 }
 
 object VacancyFakeFactory {
